Move skills challenge autos to autonomous folder
