+++ src/main/java/frc/robot/commands/autonomous/Slalom.java
@@ -1,4 +1,4 @@
-package frc.robot.commands.AutoCommands;
+package frc.robot.commands.autonomous;
 
 import com.kauailabs.navx.frc.AHRS;
 import edu.wpi.first.wpilibj.geometry.Pose2d;
@@ -9,6 +9,8 @@ import edu.wpi.first.wpilibj.trajectory.TrajectoryConfig;
 import edu.wpi.first.wpilibj.trajectory.TrajectoryGenerator;
 import edu.wpi.first.wpilibj.trajectory.constraint.CentripetalAccelerationConstraint;
 import edu.wpi.first.wpilibj.trajectory.constraint.DifferentialDriveVoltageConstraint;
+import edu.wpi.first.wpilibj.trajectory.constraint.MaxVelocityConstraint;
+import edu.wpi.first.wpilibj.trajectory.constraint.RectangularRegionConstraint;
 import edu.wpi.first.wpilibj2.command.InstantCommand;
 import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
 import frc.robot.Constants;
@@ -20,7 +22,7 @@ import frc.robot.subsystems.Limelight;
 import frc.robot.subsystems.ShooterSubsystem;
 import java.util.*;
 
-public class BarrelRacing extends SequentialCommandGroup {
+public class Slalom extends SequentialCommandGroup {
         // shoot 3 balls and move backwards to pick up 3 more and shoot new 3 balls
         private final IntakeSubsystem intakeSubsystem;
                         // .addConstraint(new RectangularRegionConstraint(new Translation2d(3.27, -3.5), new
@@ -30,7 +32,7 @@ public class BarrelRacing extends SequentialCommandGroup {
                         // .addConstraint(new RectangularRegionConstraint(new Translation2d(5.677, -1.969), new
                         // Translation2d(6.883, -0.991), new MaxVelocityConstraint(0.5)))
                         // .addConstraint(new CentripetalAccelerationConstraint(10));
-        public BarrelRacing(
+        public Slalom(
             DriveSubsystem driveSubsystem,
             Odometry odometry,
             Limelight limelight,
@@ -46,40 +48,40 @@ public class BarrelRacing extends SequentialCommandGroup {
                                         DriveSubsystem.KINEMATICS,
                                         Constants.MAX_GENERATION_VOLTAGE))
                         .addConstraint(
-                                new CentripetalAccelerationConstraint(1)
+                                new CentripetalAccelerationConstraint(2)
+                        )
+                        .addConstraint(
+                                new RectangularRegionConstraint(
+                                        new Translation2d(6.743, -3.874),
+                                        new Translation2d(8.699, -2.07),
+                                        new CentripetalAccelerationConstraint(1)        
+                                )        
                         );
-                var barrelRacing = TrajectoryGenerator.generateTrajectory(
-                        new Pose2d(1.5244232745452, -2.14663685599222, new Rotation2d(0)),
+
+                var slalom = TrajectoryGenerator.generateTrajectory(
+                        new Pose2d(1.53712151090126, -3.8735970004166, new Rotation2d(0)),
                         List.of(
-                                new Translation2d(2.85773809193167, -2.1212403832801),
-                                new Translation2d(3.79740758228023, -2.31171392862102),
-                                new Translation2d(4.34343174559089, -3.03551340091653),
-                                new Translation2d(4.10216525482571, -3.53074461880294),
-                                new Translation2d(3.55614109151506, -3.47995167337869),
-                                new Translation2d(3.27677989168171, -3.13709929176503),
-                                new Translation2d(3.61963227329537, -2.6164716011665),
-                                new Translation2d(4.88945590890154, -2.33711040133314),
-                                new Translation2d(6.01959894459102, -2.13393861963616),
-                                new Translation2d(6.50213192612137, -1.98155978336342),
-                                new Translation2d(6.76879488959866, -1.4228373836967),
-                                new Translation2d(6.24816719900013, -0.991097347590612),
-                                new Translation2d(5.62595361755311, -1.32125149284821),
-                                new Translation2d(5.67674656297736, -1.96886154700736),
-                                new Translation2d(6.29896014442438, -2.9466257464241),
-                                new Translation2d(7.09894903485626, -3.47995167337869),
-                                new Translation2d(8.05131676156089, -3.39106401888626),
-                                new Translation2d(8.22909207054575, -2.74345396472712),
-                                new Translation2d(7.64497319816692, -2.32441216497708),
-                                new Translation2d(6.07039189001527, -2.31171392862102),
-                                new Translation2d(4.48311234550756, -2.05774920149979)
+                                new Translation2d(2.43869629218164, -2.76885043743924),
+                                new Translation2d(3.50534814609082, -2.41329981946951),
+                                new Translation2d(4.99104179975003, -2.32441216497708),
+                                new Translation2d(6.47673545340924, -2.57837689209831),
+                                new Translation2d(6.99736314400777, -3.47995167337869),
+                                new Translation2d(7.6830679072351, -3.86089876406054),
+                                new Translation2d(8.43226385224274, -2.95932398278016),
+                                new Translation2d(7.7846537980836, -2.05774920149979),
+                                new Translation2d(6.98466490765171, -2.51488571031801),
+                                new Translation2d(6.74339841688654, -3.39106401888626),
+                                new Translation2d(5.91801305374253, -3.8735970004166),
+                                new Translation2d(3.8101058186363, -3.86089876406054),
+                                new Translation2d(2.33711040133314, -3.09900458269684)
                         ),
-                        new Pose2d(1.49902680183307, -2.15933509234828, new Rotation2d(3.141592654)),
+                        new Pose2d(1.34664796556033, -2.1847315650604, new Rotation2d(3.141592654)),
                         forwardConfig
                 );
 
                 addCommands(
                         new InstantCommand(() -> navx.reset()),
-                        new FollowTrajectoryCommand(barrelRacing, odometry, driveSubsystem)
+                        new FollowTrajectoryCommand(slalom, odometry, driveSubsystem)
                 );
         }
